refactor: remove unnecessary dto, move GenerateImageRequestDto to ImageProps
